Update mobile app to use shared webapi models (#900)
