+++ apps/mobile-app/components/EmailCard.tsx
@@ -2,8 +2,8 @@ import { router } from 'expo-router';
 import { useEffect, useState } from 'react';
 import { StyleSheet, View, TouchableOpacity } from 'react-native';
 
+import type { MailboxEmail } from '@/utils/shared/models';
 import { Credential } from '@/utils/types/Credential';
-import { MailboxEmail } from '@/utils/types/webapi/MailboxEmail';
 
 import { useColors } from '@/hooks/useColorScheme';
