+++ apps/mobile-app/context/DbContext.tsx
@@ -1,9 +1,9 @@
 import React, { createContext, useContext, useState, useEffect, useCallback, useMemo } from 'react';
 
+import type { VaultResponse } from '@/utils/shared/models';
 import SqliteClient from '@/utils/SqliteClient';
 import { EncryptionKeyDerivationParams } from '@/utils/types/messaging/EncryptionKeyDerivationParams';
 import { VaultMetadata } from '@/utils/types/messaging/VaultMetadata';
-import { VaultResponse } from '@/utils/types/webapi/VaultResponse';
 
 import NativeVaultManager from '@/specs/NativeVaultManager';
