+++ apps/mobile-app/utils/EncryptionUtility.tsx
@@ -3,9 +3,9 @@ import { Buffer } from 'buffer';
 import AesGcmCrypto from 'react-native-aes-gcm-crypto';
 import argon2 from 'react-native-argon2';
 
+import type { Email, MailboxEmail } from '@/utils/shared/models';
+
 import { EncryptionKey } from './types/EncryptionKey';
-import { Email } from './types/webapi/Email';
-import { MailboxEmail } from './types/webapi/MailboxEmail';
 
 /**
  * Utility class for encryption operations including:
