+++ apps/mobile-app/utils/shared/models/README.md
@@ -1,0 +1,9 @@
+# ⚠️ Auto-Generated Files
+
+This folder contains the output of the shared `models` module from the `/shared` directory in the AliasVault project.
+
+**Do not edit any of these files manually.**
+
+To make changes:
+1. Update the source files in the `/shared/models/src` directory
+2. Run the `build.sh` script in the module directory to regenerate the outputs and copy them here.
