--- apps/mobile-app/utils/types/webapi/Attachment.ts
@@ -1,19 +1,0 @@
-/**
- * Email attachment type.
- */
-export type Attachment = {
-    /** The ID of the attachment */
-    id: number;
-
-    /** The ID of the email the attachment belongs to */
-    emailId: number;
-
-    /** The filename of the attachment */
-    filename: string;
-
-    /** The MIME type of the attachment */
-    mimeType: string;
-
-    /** The size of the attachment in bytes */
-    filesize: number;
-}
