--- apps/mobile-app/utils/types/webapi/BadRequestResponse.ts
@@ -1,9 +1,0 @@
-type BadRequestResponse = {
-  type: string;
-  title: string;
-  status: number;
-  errors: Record<string, string[]>;
-  traceId: string;
-};
-
-export default BadRequestResponse;
