--- apps/mobile-app/utils/types/webapi/DeleteAccountInitiate.ts
@@ -1,16 +1,0 @@
-/**
- * Represents a delete account initiate response.
- */
-export type DeleteAccountInitiateRequest = {
-  username: string;
-};
-
-/**
- * Represents a delete account initiate response.
- */
-export type DeleteAccountInitiateResponse = {
-  salt: string;
-  serverEphemeral: string;
-  encryptionType: string;
-  encryptionSettings: string;
-};
