--- apps/mobile-app/utils/types/webapi/DeleteAccountRequest.ts
@@ -1,8 +1,0 @@
-/**
- * Represents a delete account request.
- */
-export type DeleteAccountRequest = {
-  username: string;
-  clientPublicEphemeral: string;
-  clientSessionProof: string;
-};
