--- apps/mobile-app/utils/types/webapi/FaviconExtractModel.ts
@@ -1,3 +1,0 @@
-export type FaviconExtractModel = {
-    image: string | null;
-}
