--- apps/mobile-app/utils/types/webapi/Login.ts
@@ -1,16 +1,0 @@
-/**
- * Login request type.
- */
-export type LoginRequest = {
-    username: string;
-}
-
-/**
- * Login response type.
- */
-export type LoginResponse = {
-    salt: string;
-    serverEphemeral: string;
-    encryptionType: string;
-    encryptionSettings: string;
-}
