--- apps/mobile-app/utils/types/webapi/PasswordChangeInitiateResponse.ts
@@ -1,9 +1,0 @@
-/**
- * Represents a password change initiate response.
- */
-export type PasswordChangeInitiateResponse = {
-  salt: string;
-  serverEphemeral: string;
-  encryptionType: string;
-  encryptionSettings: string;
-};
