--- apps/mobile-app/utils/types/webapi/StatusResponse.ts
@@ -1,8 +1,0 @@
-/**
- * Status response type.
- */
-export type StatusResponse = {
-  clientVersionSupported: boolean;
-  serverVersion: string;
-  vaultRevision: number;
-}
