--- apps/mobile-app/utils/types/webapi/VaultPostResponse.ts
@@ -1,7 +1,0 @@
-/**
- * Vault post response type returned after uploading a new vault to the server.
- */
-export type VaultPostResponse = {
-  status: number;
-  newRevisionNumber: number;
-}
