--- apps/mobile-app/utils/types/webapi/VaultResponse.ts
@@ -1,9 +1,0 @@
-import { Vault } from "./Vault";
-
-/**
- * Vault response type.
- */
-export type VaultResponse = {
-  status: number;
-  vault: Vault;
-}
