+++ shared/models/build.sh
@@ -6,6 +6,7 @@ set -u  # Treat unset variables as errors
 # Define output targets for models
 TARGETS=(
   "../../apps/browser-extension/src/utils/shared/models"
+  "../../apps/mobile-app/utils/shared/models"
 )
 
 # Build and distribute models
